Update base for Update on "Enhancement for smart serialization for out schemas"
Differential Revision: [D30415255](https://our.internmc.facebook.com/intern/diff/D30415255/)

Differential Revision: [D30415255](https://our.internmc.facebook.com/intern/diff/D30415255)

[ghstack-poisoned]

+++ torch/cpu/amp/autocast_mode.py
@@ -5,5 +5,5 @@ class autocast(torch.autocast_mode.autocast):
     See :class:`torch.autocast`.
     ``torch.cpu.amp.autocast(args...)`` is equivalent to ``torch.autocast("cpu", args...)``
     """
-    def __init__(self, enabled=True, dtype=torch.float16):
-        super().__init__("cpu", enabled=enabled, dtype=dtype)
+    def __init__(self, enabled=True, fast_dtype=torch.float16):
+        super().__init__("cpu", enabled=enabled, fast_dtype=fast_dtype)
